% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AminoAcidCounter.R
\name{AAcountsPlot}
\alias{AAcountsPlot}
\title{Plot Amino Acid Counts}
\usage{
AAcountsPlot(seq)
}
\arguments{
\item{seq}{An amino acid sequence}
}
\value{
Amino Acid Counts plot
}
\description{
Plot Amino Acid Counts
}
\examples{
sequence <- "ATAACGTCGTGCAAATCGGCCA"
aaCountsPlot <- AAcountsPlot(sequence)
aaCountsPlot
}
