% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonCounter.R
\name{CodonCounter}
\alias{CodonCounter}
\title{Title A codon number function}
\usage{
CodonCounter(my_seq, start = 1)
}
\arguments{
\item{my_seq}{The input sequence}

\item{start}{The first nucleotide of the sequence}
}
\value{
The output is the number of the codons this sequence comprises of
}
\description{
Title A codon number function
}
\examples{
 my_seq <- "ATGGTACCTGA"
 codons <- CodonCounter(my_seq)
 print(codons)
}
