% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodonFinder.R
\name{CodonFinder}
\alias{CodonFinder}
\title{Title A codon Finder function}
\usage{
CodonFinder(my_seq, start = 1)
}
\arguments{
\item{my_seq}{The input sequence}

\item{start}{The first nucleotide of the sequence}
}
\value{
The output is the number of the codons this sequence comprises of
}
\description{
Title A codon Finder function
}
\examples{
 my_seq <- "ATGGTACCTGA"
 codons <- CodonFinder(my_seq)
 print(codons)
}
