% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CodonTable}
\alias{CodonTable}
\title{Decription of data}
\format{
An object of class \code{character} of length 64.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
CodonTable
}
\description{
This variable is the Codon Table with all possbile codons and the aninoacids that they translate to
}
\keyword{datasets}
