% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_function.R
\name{Create_DNA}
\alias{Create_DNA}
\title{Create_DNA}
\usage{
Create_DNA(n)
}
\arguments{
\item{n|}{This indicates the length of the sequence}
}
\value{
the sequence as a string
}
\description{
Create_DNA
}
