% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcribe_dna.R
\name{dna_to_RNA}
\alias{dna_to_RNA}
\title{Transcribe a DNA sequence into a RNA sequence}
\usage{
dna_to_RNA(dna_sequence)
}
\arguments{
\item{dna_sequence}{A character string representing a DNA sequence}
}
\value{
A character string respresenting the transcribed RNA sequence
}
\description{
Transcribe a DNA sequence into a RNA sequence
}
