% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mRNA_to_codon.R
\name{mRNA_to_codon}
\alias{mRNA_to_codon}
\title{mRNA_to_codon}
\usage{
mRNA_to_codon(mRNA, start = 1)
}
\arguments{
\item{mRNA}{mRNA to transform}

\item{start}{When to start the transformation. Default set to 1}
}
\value{
Replaces mRNA with codon in string
}
\description{
mRNA_to_codon
}
\examples{
mRNA <- "AUGUCG"
mRNA_to_codon(mRNA)
}
