% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_chain.R
\name{ProChain}
\alias{ProChain}
\title{Protein Sequence}
\usage{
ProChain(codons)
}
\arguments{
\item{codons}{the triplet of noucleotides}
}
\value{
the amino acid sequence of given codons
}
\description{
Protein Sequence
}
\examples{
ProChain(c("UUA","AAA","CAG"))

}
