% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplDna.R
\name{ReplDna}
\alias{ReplDna}
\title{ReplDna}
\usage{
ReplDna(sequ)
}
\arguments{
\item{sequ}{The DNA sequence provided by the user}
}
\value{
Takes a given DNA sequence and returns the complementary strand
}
\description{
ReplDna
}
\examples{
my_seq <- "ATCGATT"
my_compl <- ReplDna(my_seq)
my_compl
}
