% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevTrans.R
\name{RevTrans}
\alias{RevTrans}
\title{RevTrans}
\usage{
RevTrans(sequ)
}
\arguments{
\item{sequ}{The RNA sequence provided by the user}
}
\value{
Takes a given RNA sequence and returns the complementary DNA sequence
after reverse transcription.
}
\description{
RevTrans
}
\examples{
my_seq <- "AUCGAUU"
my_compl <- RevTrans(my_seq)
my_compl
}
