% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcr.R
\name{Transcr}
\alias{Transcr}
\title{Converts the DNA sequence to an RNA sequence}
\usage{
Transcr(sequence)
}
\arguments{
\item{sequence}{the sequence provided by the user}
}
\value{
Takes a given DNA sequence and converts it to an RNA sequence,
changing T to U
}
\description{
Converts the DNA sequence to an RNA sequence
}
