% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation.R
\name{translation}
\alias{translation}
\title{Translation}
\usage{
translation(codons)
}
\arguments{
\item{codons, }{takes a string of codons}
}
\value{
a string of aminoacids which have been translated from the codons parameter
}
\description{
Translates a list of codons into teir corresponding amino acids
}
\examples{
codons_example <- c("AUG", "UCU",  "AAG")
translation(codons_example)
}
