% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqLen.R
\name{SeqLen}
\alias{SeqLen}
\title{SeqLen}
\usage{
SeqLen(length)
}
\arguments{
\item{length}{the user decides the sequence length, provides a number}
}
\value{
creates a nucleotide sequence of given length
}
\description{
SeqLen
}
\examples{
my_length <- 10
my_seq <- SeqLen(my_length)
my_seq
}
