% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transcr.R
\name{Transcr}
\alias{Transcr}
\title{Trasncr}
\usage{
Transcr(sequence)
}
\arguments{
\item{sequence}{The sequence provided by the user}
}
\value{
Takes a given DNA sequence and converts it to an RNA sequence,
changing T to U
}
\description{
Trasncr
}
\examples{
my_seq <- "ATCGTA"
my_rna <- Transcr(my_seq)
my_rna
}
